+++ examples/pub/lib/detail.g.dart
@@ -32,8 +32,6 @@ class _SystemHash {
   }
 }
 
-typedef FetchPackageDetailsRef = AutoDisposeFutureProviderRef<Package>;
-
 /// See also [fetchPackageDetails].
 @ProviderFor(fetchPackageDetails)
 const fetchPackageDetailsProvider = FetchPackageDetailsFamily();
@@ -80,10 +78,10 @@ class FetchPackageDetailsFamily extends Family<AsyncValue<Package>> {
 class FetchPackageDetailsProvider extends AutoDisposeFutureProvider<Package> {
   /// See also [fetchPackageDetails].
   FetchPackageDetailsProvider({
-    required this.packageName,
-  }) : super.internal(
+    required String packageName,
+  }) : this._internal(
           (ref) => fetchPackageDetails(
-            ref,
+            ref as FetchPackageDetailsRef,
             packageName: packageName,
           ),
           from: fetchPackageDetailsProvider,
@@ -95,9 +93,43 @@ class FetchPackageDetailsProvider extends AutoDisposeFutureProvider<Package> {
           dependencies: FetchPackageDetailsFamily._dependencies,
           allTransitiveDependencies:
               FetchPackageDetailsFamily._allTransitiveDependencies,
+          packageName: packageName,
         );
 
+  FetchPackageDetailsProvider._internal(
+    super._createNotifier, {
+    required super.name,
+    required super.dependencies,
+    required super.allTransitiveDependencies,
+    required super.debugGetCreateSourceHash,
+    required super.from,
+    required this.packageName,
+  }) : super.internal();
+
   final String packageName;
+
+  @override
+  Override overrideWith(
+    FutureOr<Package> Function(FetchPackageDetailsRef provider) create,
+  ) {
+    return ProviderOverride(
+      origin: this,
+      override: FetchPackageDetailsProvider._internal(
+        (ref) => create(ref as FetchPackageDetailsRef),
+        from: from,
+        name: null,
+        dependencies: null,
+        allTransitiveDependencies: null,
+        debugGetCreateSourceHash: null,
+        packageName: packageName,
+      ),
+    );
+  }
+
+  @override
+  AutoDisposeFutureProviderElement<Package> createElement() {
+    return _FetchPackageDetailsProviderElement(this);
+  }
 
   @override
   bool operator ==(Object other) {
@@ -112,6 +144,20 @@ class FetchPackageDetailsProvider extends AutoDisposeFutureProvider<Package> {
 
     return _SystemHash.finish(hash);
   }
+}
+
+mixin FetchPackageDetailsRef on AutoDisposeFutureProviderRef<Package> {
+  /// The parameter `packageName` of this provider.
+  String get packageName;
+}
+
+class _FetchPackageDetailsProviderElement
+    extends AutoDisposeFutureProviderElement<Package>
+    with FetchPackageDetailsRef {
+  _FetchPackageDetailsProviderElement(super.provider);
+
+  @override
+  String get packageName => (origin as FetchPackageDetailsProvider).packageName;
 }
 
 String _$likedPackagesHash() => r'304a4def167e245812638cba776e8d5eb66d8844';
@@ -238,8 +284,8 @@ class PackageMetricsProvider extends AutoDisposeAsyncNotifierProviderImpl<
   ///
   /// Copied from [PackageMetrics].
   PackageMetricsProvider({
-    required this.packageName,
-  }) : super.internal(
+    required String packageName,
+  }) : this._internal(
           () => PackageMetrics()..packageName = packageName,
           from: packageMetricsProvider,
           name: r'packageMetricsProvider',
@@ -250,9 +296,51 @@ class PackageMetricsProvider extends AutoDisposeAsyncNotifierProviderImpl<
           dependencies: PackageMetricsFamily._dependencies,
           allTransitiveDependencies:
               PackageMetricsFamily._allTransitiveDependencies,
+          packageName: packageName,
         );
 
+  PackageMetricsProvider._internal(
+    super._createNotifier, {
+    required super.name,
+    required super.dependencies,
+    required super.allTransitiveDependencies,
+    required super.debugGetCreateSourceHash,
+    required super.from,
+    required this.packageName,
+  }) : super.internal();
+
   final String packageName;
+
+  @override
+  Future<PackageMetricsScore> runNotifierBuild(
+    covariant PackageMetrics notifier,
+  ) {
+    return notifier.build(
+      packageName: packageName,
+    );
+  }
+
+  @override
+  Override overrideWith(PackageMetrics Function() create) {
+    return ProviderOverride(
+      origin: this,
+      override: PackageMetricsProvider._internal(
+        () => create()..packageName = packageName,
+        from: from,
+        name: null,
+        dependencies: null,
+        allTransitiveDependencies: null,
+        debugGetCreateSourceHash: null,
+        packageName: packageName,
+      ),
+    );
+  }
+
+  @override
+  AutoDisposeAsyncNotifierProviderElement<PackageMetrics, PackageMetricsScore>
+      createElement() {
+    return _PackageMetricsProviderElement(this);
+  }
 
   @override
   bool operator ==(Object other) {
@@ -266,15 +354,21 @@ class PackageMetricsProvider extends AutoDisposeAsyncNotifierProviderImpl<
 
     return _SystemHash.finish(hash);
   }
+}
+
+mixin PackageMetricsRef
+    on AutoDisposeAsyncNotifierProviderRef<PackageMetricsScore> {
+  /// The parameter `packageName` of this provider.
+  String get packageName;
+}
+
+class _PackageMetricsProviderElement
+    extends AutoDisposeAsyncNotifierProviderElement<PackageMetrics,
+        PackageMetricsScore> with PackageMetricsRef {
+  _PackageMetricsProviderElement(super.provider);
 
   @override
-  Future<PackageMetricsScore> runNotifierBuild(
-    covariant PackageMetrics notifier,
-  ) {
-    return notifier.build(
-      packageName: packageName,
-    );
-  }
+  String get packageName => (origin as PackageMetricsProvider).packageName;
 }
 // ignore_for_file: type=lint
 // ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member