+++ packages/riverpod/lib/src/async_notifier/auto_dispose_family.dart
@@ -70,7 +70,7 @@ class AutoDisposeFamilyAsyncNotifierProviderImpl<
 
   @override
   AutoDisposeAsyncNotifierProviderElement<NotifierT, T> createElement() {
-    return AutoDisposeAsyncNotifierProviderElement._(this);
+    return AutoDisposeAsyncNotifierProviderElement(this);
   }
 
   @override