+++ packages/riverpod_generator/lib/src/templates/parameters.dart
@@ -3,19 +3,33 @@ import 'package:analyzer/dart/element/element.dart';
 String buildParamDefinitionQuery(
   List<ParameterElement> parameters, {
   bool asThisParameter = false,
+  bool asSuperParameter = false,
+  bool writeBrackets = true,
+  bool asRequiredNamed = false,
 }) {
-  final requiredPositionals =
-      parameters.where((element) => element.isRequiredPositional);
-  final optionalPositionals =
-      parameters.where((element) => element.isOptionalPositional).toList();
-  final named = parameters.where((element) => element.isNamed).toList();
+  assert(
+    !asThisParameter || !asSuperParameter,
+    'Cannot enable both asThisParameter and asSuperParameter',
+  );
+
+  final requiredPositionals = parameters
+      .where((element) => element.isRequiredPositional && !asRequiredNamed)
+      .toList();
+  final optionalPositionals = parameters
+      .where((element) => element.isOptionalPositional && !asRequiredNamed)
+      .toList();
+  final named = parameters
+      .where((element) => element.isNamed || asRequiredNamed)
+      .toList();
 
   final buffer = StringBuffer();
   String encodeParameter(ParameterElement e) {
-    final leading = e.isRequiredNamed ? 'required ' : '';
-    final trailing =
-        e.defaultValueCode != null ? '= ${e.defaultValueCode}' : '';
+    final leading = e.isRequiredNamed || asRequiredNamed ? 'required ' : '';
+    final trailing = e.defaultValueCode != null && !asRequiredNamed
+        ? '= ${e.defaultValueCode}'
+        : '';
     if (asThisParameter) return '${leading}this.${e.name}$trailing';
+    if (asSuperParameter) return '${leading}super.${e.name}$trailing';
     return '$leading${e.type} ${e.name}$trailing';
   }
 
@@ -23,18 +37,16 @@ String buildParamDefinitionQuery(
     requiredPositionals.map(encodeParameter).expand((e) => [e, ',']),
   );
   if (optionalPositionals.isNotEmpty) {
-    buffer
-      ..write('[')
-      ..writeAll(
-        optionalPositionals.map(encodeParameter).expand((e) => [e, ',']),
-      )
-      ..write(']');
+    if (writeBrackets) buffer.write('[');
+    buffer.writeAll(
+      optionalPositionals.map(encodeParameter).expand((e) => [e, ',']),
+    );
+    if (writeBrackets) buffer.write(']');
   }
   if (named.isNotEmpty) {
-    buffer
-      ..write('{')
-      ..writeAll(named.map(encodeParameter).expand((e) => [e, ',']))
-      ..write('}');
+    if (writeBrackets) buffer.write('{');
+    buffer.writeAll(named.map(encodeParameter).expand((e) => [e, ',']));
+    if (writeBrackets) buffer.write('}');
   }
 
   return buffer.toString();