+++ packages/riverpod_generator/test/async_notifier_test.dart
@@ -27,9 +27,44 @@ void main() {
     expect(familyClassProvider(42, third: .42).name, 'familyClassProvider');
   });
 
+  test('Supports overriding non-family notifiers', () {
+    final container = createContainer(
+      overrides: [
+        publicClassProvider.overrideWith(() => PublicClass('Hello world')),
+      ],
+    );
+
+    final notifier = container.read(publicClassProvider.notifier);
+    expect(notifier.param, 'Hello world');
+
+    expect(notifier.ref, isNotNull);
+    expect(notifier.state, isNotNull);
+  });
+
+  test('Supports overriding family notifiers', () {
+    final container = createContainer(
+      overrides: [
+        familyClassProvider(42, third: .42)
+            .overrideWith(() => FamilyClass('Hello world')),
+      ],
+    );
+
+    final notifier =
+        container.read(familyClassProvider(42, third: .42).notifier);
+    expect(notifier.param, 'Hello world');
+    expect(notifier.first, 42);
+    expect(notifier.second, null);
+    expect(notifier.third, .42);
+    expect(notifier.fourth, true);
+    expect(notifier.fifth, null);
+
+    expect(notifier.ref, isNotNull);
+    expect(notifier.state, isNotNull);
+  });
+
   test(
       'Creates a NotifierProvider.family<T> if @riverpod is used on a synchronous function with parameters',
-      () {
+      () async {
     final container = createContainer();
 
     const FamilyClassFamily family = familyClassProvider;
@@ -74,7 +109,7 @@ void main() {
       second: 'x42',
       third: .42,
       fourth: false,
-      fifth: ['x42'],
+      fifth: const ['x42'],
     );
     // ignore: invalid_use_of_internal_member
     final AutoDisposeAsyncNotifierProviderImpl<FamilyClass, String>
@@ -84,7 +119,19 @@ void main() {
     expect(provider.second, 'x42');
     expect(provider.third, .42);
     expect(provider.fourth, false);
-    expect(provider.fifth, ['x42']);
+    expect(provider.fifth, same(const ['x42']));
+
+    final sub = container.listen(
+      familyClassProvider(
+        42,
+        second: 'x42',
+        third: .42,
+        fourth: false,
+        fifth: const ['x42'],
+      ).future,
+      (previous, next) {},
+    );
+    await sub.read();
 
     final AsyncValue<String> result = container.read(
       familyClassProvider(
@@ -92,7 +139,7 @@ void main() {
         second: 'x42',
         third: .42,
         fourth: false,
-        fifth: ['x42'],
+        fifth: const ['x42'],
       ),
     );