+++ packages/riverpod_generator/test/integration/sync.g.dart
@@ -57,8 +57,6 @@ class _SystemHash {
   }
 }
 
-typedef RawFamilyFutureRef = AutoDisposeProviderRef<Future<String>>;
-
 /// See also [rawFamilyFuture].
 @ProviderFor(rawFamilyFuture)
 const rawFamilyFutureProvider = RawFamilyFutureFamily();
@@ -105,10 +103,10 @@ class RawFamilyFutureFamily extends Family<Future<String>> {
 class RawFamilyFutureProvider extends AutoDisposeProvider<Future<String>> {
   /// See also [rawFamilyFuture].
   RawFamilyFutureProvider(
-    this.id,
-  ) : super.internal(
+    int id,
+  ) : this._internal(
           (ref) => rawFamilyFuture(
-            ref,
+            ref as RawFamilyFutureRef,
             id,
           ),
           from: rawFamilyFutureProvider,
@@ -120,9 +118,43 @@ class RawFamilyFutureProvider extends AutoDisposeProvider<Future<String>> {
           dependencies: RawFamilyFutureFamily._dependencies,
           allTransitiveDependencies:
               RawFamilyFutureFamily._allTransitiveDependencies,
+          id: id,
         );
 
+  RawFamilyFutureProvider._internal(
+    super._createNotifier, {
+    required super.name,
+    required super.dependencies,
+    required super.allTransitiveDependencies,
+    required super.debugGetCreateSourceHash,
+    required super.from,
+    required this.id,
+  }) : super.internal();
+
   final int id;
+
+  @override
+  Override overrideWith(
+    Future<String> Function(RawFamilyFutureRef provider) create,
+  ) {
+    return ProviderOverride(
+      origin: this,
+      override: RawFamilyFutureProvider._internal(
+        (ref) => create(ref as RawFamilyFutureRef),
+        from: from,
+        name: null,
+        dependencies: null,
+        allTransitiveDependencies: null,
+        debugGetCreateSourceHash: null,
+        id: id,
+      ),
+    );
+  }
+
+  @override
+  AutoDisposeProviderElement<Future<String>> createElement() {
+    return _RawFamilyFutureProviderElement(this);
+  }
 
   @override
   bool operator ==(Object other) {
@@ -138,8 +170,20 @@ class RawFamilyFutureProvider extends AutoDisposeProvider<Future<String>> {
   }
 }
 
+mixin RawFamilyFutureRef on AutoDisposeProviderRef<Future<String>> {
+  /// The parameter `id` of this provider.
+  int get id;
+}
+
+class _RawFamilyFutureProviderElement
+    extends AutoDisposeProviderElement<Future<String>> with RawFamilyFutureRef {
+  _RawFamilyFutureProviderElement(super.provider);
+
+  @override
+  int get id => (origin as RawFamilyFutureProvider).id;
+}
+
 String _$rawFamilyStreamHash() => r'e778e5cfcb8ab381e2412f5c73213aaa03b93012';
-typedef RawFamilyStreamRef = AutoDisposeProviderRef<Stream<String>>;
 
 /// See also [rawFamilyStream].
 @ProviderFor(rawFamilyStream)
@@ -187,10 +231,10 @@ class RawFamilyStreamFamily extends Family<Stream<String>> {
 class RawFamilyStreamProvider extends AutoDisposeProvider<Stream<String>> {
   /// See also [rawFamilyStream].
   RawFamilyStreamProvider(
-    this.id,
-  ) : super.internal(
+    int id,
+  ) : this._internal(
           (ref) => rawFamilyStream(
-            ref,
+            ref as RawFamilyStreamRef,
             id,
           ),
           from: rawFamilyStreamProvider,
@@ -202,9 +246,43 @@ class RawFamilyStreamProvider extends AutoDisposeProvider<Stream<String>> {
           dependencies: RawFamilyStreamFamily._dependencies,
           allTransitiveDependencies:
               RawFamilyStreamFamily._allTransitiveDependencies,
+          id: id,
         );
 
+  RawFamilyStreamProvider._internal(
+    super._createNotifier, {
+    required super.name,
+    required super.dependencies,
+    required super.allTransitiveDependencies,
+    required super.debugGetCreateSourceHash,
+    required super.from,
+    required this.id,
+  }) : super.internal();
+
   final int id;
+
+  @override
+  Override overrideWith(
+    Stream<String> Function(RawFamilyStreamRef provider) create,
+  ) {
+    return ProviderOverride(
+      origin: this,
+      override: RawFamilyStreamProvider._internal(
+        (ref) => create(ref as RawFamilyStreamRef),
+        from: from,
+        name: null,
+        dependencies: null,
+        allTransitiveDependencies: null,
+        debugGetCreateSourceHash: null,
+        id: id,
+      ),
+    );
+  }
+
+  @override
+  AutoDisposeProviderElement<Stream<String>> createElement() {
+    return _RawFamilyStreamProviderElement(this);
+  }
 
   @override
   bool operator ==(Object other) {
@@ -218,6 +296,19 @@ class RawFamilyStreamProvider extends AutoDisposeProvider<Stream<String>> {
 
     return _SystemHash.finish(hash);
   }
+}
+
+mixin RawFamilyStreamRef on AutoDisposeProviderRef<Stream<String>> {
+  /// The parameter `id` of this provider.
+  int get id;
+}
+
+class _RawFamilyStreamProviderElement
+    extends AutoDisposeProviderElement<Stream<String>> with RawFamilyStreamRef {
+  _RawFamilyStreamProviderElement(super.provider);
+
+  @override
+  int get id => (origin as RawFamilyStreamProvider).id;
 }
 
 String _$publicHash() => r'138be35943899793ab085e711fe3f3d22696a3ba';
@@ -252,7 +343,6 @@ final supports$inNamesProvider = AutoDisposeProvider<String>.internal(
 
 typedef Supports$inNamesRef = AutoDisposeProviderRef<String>;
 String _$familyHash() => r'14d1ee238ca608d547630d0e222ef4c5866e9e61';
-typedef FamilyRef = AutoDisposeProviderRef<String>;
 
 /// This is some documentation
 ///
@@ -324,14 +414,14 @@ class FamilyProvider extends AutoDisposeProvider<String> {
   ///
   /// Copied from [family].
   FamilyProvider(
-    this.first, {
-    this.second,
-    required this.third,
-    this.fourth = true,
-    this.fifth,
-  }) : super.internal(
+    int first, {
+    String? second,
+    required double third,
+    bool fourth = true,
+    List<String>? fifth,
+  }) : this._internal(
           (ref) => family(
-            ref,
+            ref as FamilyRef,
             first,
             second: second,
             third: third,
@@ -346,13 +436,59 @@ class FamilyProvider extends AutoDisposeProvider<String> {
                   : _$familyHash,
           dependencies: FamilyFamily._dependencies,
           allTransitiveDependencies: FamilyFamily._allTransitiveDependencies,
+          first: first,
+          second: second,
+          third: third,
+          fourth: fourth,
+          fifth: fifth,
         );
+
+  FamilyProvider._internal(
+    super._createNotifier, {
+    required super.name,
+    required super.dependencies,
+    required super.allTransitiveDependencies,
+    required super.debugGetCreateSourceHash,
+    required super.from,
+    required this.first,
+    required this.second,
+    required this.third,
+    required this.fourth,
+    required this.fifth,
+  }) : super.internal();
 
   final int first;
   final String? second;
   final double third;
   final bool fourth;
   final List<String>? fifth;
+
+  @override
+  Override overrideWith(
+    String Function(FamilyRef provider) create,
+  ) {
+    return ProviderOverride(
+      origin: this,
+      override: FamilyProvider._internal(
+        (ref) => create(ref as FamilyRef),
+        from: from,
+        name: null,
+        dependencies: null,
+        allTransitiveDependencies: null,
+        debugGetCreateSourceHash: null,
+        first: first,
+        second: second,
+        third: third,
+        fourth: fourth,
+        fifth: fifth,
+      ),
+    );
+  }
+
+  @override
+  AutoDisposeProviderElement<String> createElement() {
+    return _FamilyProviderElement(this);
+  }
 
   @override
   bool operator ==(Object other) {
@@ -375,6 +511,39 @@ class FamilyProvider extends AutoDisposeProvider<String> {
 
     return _SystemHash.finish(hash);
   }
+}
+
+mixin FamilyRef on AutoDisposeProviderRef<String> {
+  /// The parameter `first` of this provider.
+  int get first;
+
+  /// The parameter `second` of this provider.
+  String? get second;
+
+  /// The parameter `third` of this provider.
+  double get third;
+
+  /// The parameter `fourth` of this provider.
+  bool get fourth;
+
+  /// The parameter `fifth` of this provider.
+  List<String>? get fifth;
+}
+
+class _FamilyProviderElement extends AutoDisposeProviderElement<String>
+    with FamilyRef {
+  _FamilyProviderElement(super.provider);
+
+  @override
+  int get first => (origin as FamilyProvider).first;
+  @override
+  String? get second => (origin as FamilyProvider).second;
+  @override
+  double get third => (origin as FamilyProvider).third;
+  @override
+  bool get fourth => (origin as FamilyProvider).fourth;
+  @override
+  List<String>? get fifth => (origin as FamilyProvider).fifth;
 }
 
 String _$privateHash() => r'519561bc7e88e394d7f75ca2102a5c0acc832c66';
@@ -496,8 +665,8 @@ class RawFamilyFutureClassProvider extends AutoDisposeNotifierProviderImpl<
     RawFamilyFutureClass, Future<String>> {
   /// See also [RawFamilyFutureClass].
   RawFamilyFutureClassProvider(
-    this.id,
-  ) : super.internal(
+    int id,
+  ) : this._internal(
           () => RawFamilyFutureClass()..id = id,
           from: rawFamilyFutureClassProvider,
           name: r'rawFamilyFutureClassProvider',
@@ -508,9 +677,51 @@ class RawFamilyFutureClassProvider extends AutoDisposeNotifierProviderImpl<
           dependencies: RawFamilyFutureClassFamily._dependencies,
           allTransitiveDependencies:
               RawFamilyFutureClassFamily._allTransitiveDependencies,
+          id: id,
         );
 
+  RawFamilyFutureClassProvider._internal(
+    super._createNotifier, {
+    required super.name,
+    required super.dependencies,
+    required super.allTransitiveDependencies,
+    required super.debugGetCreateSourceHash,
+    required super.from,
+    required this.id,
+  }) : super.internal();
+
   final int id;
+
+  @override
+  Future<String> runNotifierBuild(
+    covariant RawFamilyFutureClass notifier,
+  ) {
+    return notifier.build(
+      id,
+    );
+  }
+
+  @override
+  Override overrideWith(RawFamilyFutureClass Function() create) {
+    return ProviderOverride(
+      origin: this,
+      override: RawFamilyFutureClassProvider._internal(
+        () => create()..id = id,
+        from: from,
+        name: null,
+        dependencies: null,
+        allTransitiveDependencies: null,
+        debugGetCreateSourceHash: null,
+        id: id,
+      ),
+    );
+  }
+
+  @override
+  AutoDisposeNotifierProviderElement<RawFamilyFutureClass, Future<String>>
+      createElement() {
+    return _RawFamilyFutureClassProviderElement(this);
+  }
 
   @override
   bool operator ==(Object other) {
@@ -524,15 +735,21 @@ class RawFamilyFutureClassProvider extends AutoDisposeNotifierProviderImpl<
 
     return _SystemHash.finish(hash);
   }
+}
+
+mixin RawFamilyFutureClassRef
+    on AutoDisposeNotifierProviderRef<Future<String>> {
+  /// The parameter `id` of this provider.
+  int get id;
+}
+
+class _RawFamilyFutureClassProviderElement
+    extends AutoDisposeNotifierProviderElement<RawFamilyFutureClass,
+        Future<String>> with RawFamilyFutureClassRef {
+  _RawFamilyFutureClassProviderElement(super.provider);
 
   @override
-  Future<String> runNotifierBuild(
-    covariant RawFamilyFutureClass notifier,
-  ) {
-    return notifier.build(
-      id,
-    );
-  }
+  int get id => (origin as RawFamilyFutureClassProvider).id;
 }
 
 String _$rawFamilyStreamClassHash() =>
@@ -594,8 +811,8 @@ class RawFamilyStreamClassProvider extends AutoDisposeNotifierProviderImpl<
     RawFamilyStreamClass, Stream<String>> {
   /// See also [RawFamilyStreamClass].
   RawFamilyStreamClassProvider(
-    this.id,
-  ) : super.internal(
+    int id,
+  ) : this._internal(
           () => RawFamilyStreamClass()..id = id,
           from: rawFamilyStreamClassProvider,
           name: r'rawFamilyStreamClassProvider',
@@ -606,9 +823,51 @@ class RawFamilyStreamClassProvider extends AutoDisposeNotifierProviderImpl<
           dependencies: RawFamilyStreamClassFamily._dependencies,
           allTransitiveDependencies:
               RawFamilyStreamClassFamily._allTransitiveDependencies,
+          id: id,
         );
 
+  RawFamilyStreamClassProvider._internal(
+    super._createNotifier, {
+    required super.name,
+    required super.dependencies,
+    required super.allTransitiveDependencies,
+    required super.debugGetCreateSourceHash,
+    required super.from,
+    required this.id,
+  }) : super.internal();
+
   final int id;
+
+  @override
+  Stream<String> runNotifierBuild(
+    covariant RawFamilyStreamClass notifier,
+  ) {
+    return notifier.build(
+      id,
+    );
+  }
+
+  @override
+  Override overrideWith(RawFamilyStreamClass Function() create) {
+    return ProviderOverride(
+      origin: this,
+      override: RawFamilyStreamClassProvider._internal(
+        () => create()..id = id,
+        from: from,
+        name: null,
+        dependencies: null,
+        allTransitiveDependencies: null,
+        debugGetCreateSourceHash: null,
+        id: id,
+      ),
+    );
+  }
+
+  @override
+  AutoDisposeNotifierProviderElement<RawFamilyStreamClass, Stream<String>>
+      createElement() {
+    return _RawFamilyStreamClassProviderElement(this);
+  }
 
   @override
   bool operator ==(Object other) {
@@ -622,18 +881,24 @@ class RawFamilyStreamClassProvider extends AutoDisposeNotifierProviderImpl<
 
     return _SystemHash.finish(hash);
   }
-
-  @override
-  Stream<String> runNotifierBuild(
-    covariant RawFamilyStreamClass notifier,
-  ) {
-    return notifier.build(
-      id,
-    );
-  }
 }
 
-String _$publicClassHash() => r'f04884c039e6200ad3537feeecfc6e83828b5eb5';
+mixin RawFamilyStreamClassRef
+    on AutoDisposeNotifierProviderRef<Stream<String>> {
+  /// The parameter `id` of this provider.
+  int get id;
+}
+
+class _RawFamilyStreamClassProviderElement
+    extends AutoDisposeNotifierProviderElement<RawFamilyStreamClass,
+        Stream<String>> with RawFamilyStreamClassRef {
+  _RawFamilyStreamClassProviderElement(super.provider);
+
+  @override
+  int get id => (origin as RawFamilyStreamClassProvider).id;
+}
+
+String _$publicClassHash() => r'c8e7eec9e202acf8394e02496857cbe49405bf62';
 
 /// This is some documentation
 ///
@@ -665,7 +930,7 @@ final _privateClassProvider =
 );
 
 typedef _$PrivateClass = AutoDisposeNotifier<String>;
-String _$familyClassHash() => r'7dd0013dba8f45e82e8e39fbb2635e5a7f4b9cac';
+String _$familyClassHash() => r'01e3b9cb4d6d0bf12a2284761b1a11819d97d249';
 
 abstract class _$FamilyClass extends BuildlessAutoDisposeNotifier<String> {
   late final int first;
@@ -754,12 +1019,12 @@ class FamilyClassProvider
   ///
   /// Copied from [FamilyClass].
   FamilyClassProvider(
-    this.first, {
-    this.second,
-    required this.third,
-    this.fourth = true,
-    this.fifth,
-  }) : super.internal(
+    int first, {
+    String? second,
+    required double third,
+    bool fourth = true,
+    List<String>? fifth,
+  }) : this._internal(
           () => FamilyClass()
             ..first = first
             ..second = second
@@ -775,13 +1040,75 @@ class FamilyClassProvider
           dependencies: FamilyClassFamily._dependencies,
           allTransitiveDependencies:
               FamilyClassFamily._allTransitiveDependencies,
+          first: first,
+          second: second,
+          third: third,
+          fourth: fourth,
+          fifth: fifth,
         );
+
+  FamilyClassProvider._internal(
+    super._createNotifier, {
+    required super.name,
+    required super.dependencies,
+    required super.allTransitiveDependencies,
+    required super.debugGetCreateSourceHash,
+    required super.from,
+    required this.first,
+    required this.second,
+    required this.third,
+    required this.fourth,
+    required this.fifth,
+  }) : super.internal();
 
   final int first;
   final String? second;
   final double third;
   final bool fourth;
   final List<String>? fifth;
+
+  @override
+  String runNotifierBuild(
+    covariant FamilyClass notifier,
+  ) {
+    return notifier.build(
+      first,
+      second: second,
+      third: third,
+      fourth: fourth,
+      fifth: fifth,
+    );
+  }
+
+  @override
+  Override overrideWith(FamilyClass Function() create) {
+    return ProviderOverride(
+      origin: this,
+      override: FamilyClassProvider._internal(
+        () => create()
+          ..first = first
+          ..second = second
+          ..third = third
+          ..fourth = fourth
+          ..fifth = fifth,
+        from: from,
+        name: null,
+        dependencies: null,
+        allTransitiveDependencies: null,
+        debugGetCreateSourceHash: null,
+        first: first,
+        second: second,
+        third: third,
+        fourth: fourth,
+        fifth: fifth,
+      ),
+    );
+  }
+
+  @override
+  AutoDisposeNotifierProviderElement<FamilyClass, String> createElement() {
+    return _FamilyClassProviderElement(this);
+  }
 
   @override
   bool operator ==(Object other) {
@@ -804,19 +1131,40 @@ class FamilyClassProvider
 
     return _SystemHash.finish(hash);
   }
+}
+
+mixin FamilyClassRef on AutoDisposeNotifierProviderRef<String> {
+  /// The parameter `first` of this provider.
+  int get first;
+
+  /// The parameter `second` of this provider.
+  String? get second;
+
+  /// The parameter `third` of this provider.
+  double get third;
+
+  /// The parameter `fourth` of this provider.
+  bool get fourth;
+
+  /// The parameter `fifth` of this provider.
+  List<String>? get fifth;
+}
+
+class _FamilyClassProviderElement
+    extends AutoDisposeNotifierProviderElement<FamilyClass, String>
+    with FamilyClassRef {
+  _FamilyClassProviderElement(super.provider);
 
   @override
-  String runNotifierBuild(
-    covariant FamilyClass notifier,
-  ) {
-    return notifier.build(
-      first,
-      second: second,
-      third: third,
-      fourth: fourth,
-      fifth: fifth,
-    );
-  }
+  int get first => (origin as FamilyClassProvider).first;
+  @override
+  String? get second => (origin as FamilyClassProvider).second;
+  @override
+  double get third => (origin as FamilyClassProvider).third;
+  @override
+  bool get fourth => (origin as FamilyClassProvider).fourth;
+  @override
+  List<String>? get fifth => (origin as FamilyClassProvider).fifth;
 }
 
 String _$supports$InClassNameHash() =>