+++ packages/hooks_riverpod/CHANGELOG.md
@@ -1,3 +1,18 @@
+## 2.3.10 - 2023-08-28
+
+- `flutter_riverpod` upgraded to `2.3.10`
+- `riverpod` upgraded to `2.3.10`
+
+## 2.3.9 - 2023-08-28
+
+- Fix some exceptions causing Flutter asking to demangle stacktraces (#1874)
+- Fix out of date `AsyncValue` docs.
+
+## 2.3.8 - 2023-08-16
+
+- Added support for state_notifier 1.0.0
+- `Notifier.state` is now accessible inside tests
+
 ## 2.3.7 - 2023-07-25
 
 - Support flutter_hooks 0.20.0