+++ packages/riverpod_annotation/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 2.1.4 - 2023-08-28
+
+- `riverpod` upgraded to `2.3.10`
+
+## 2.1.3 - 2023-08-28
+
+- `riverpod` upgraded to `2.3.8`
+
+## 2.1.2 - 2023-08-16
+
+- `riverpod` upgraded to `2.3.7`
+
 ## 2.1.1 - 2023-04-24
 
 - `riverpod` upgraded to `2.3.6`